Added api for spectrum io
