+++ spectrum_io/search_result/__init__.py
@@ -1,4 +1,2 @@
 """Initialize seach result."""
-from .mascot import Mascot
-from .maxquant import MaxQuant
-from .msfragger import MSFragger
+from .process import *
